refactor: move helper methods implementation inside `ProgramCall` trait
